Allow worker to be disposed when golang proxy is reclaimed
This change sets a finalizer on the golang "worker" object, which allows
it to send a worker_dispose signal to the C++ side of things which in
turn allows the C++ code to Dispose() of the V8 Isolate, as well as
"delete" the C++ wrapping struct for the worker.

This effectively means that whenever a worker is removed by the go GC,
it gets clean up properly.

I wrote a UT to create 10000 workers in a loop, and without the
finalizer, noticed that memory usage went up to 3 GBs. However with the
finalizer code, the test ran much faster and general performance was up,
along with memory usage being capped at only 8 MBs.

The process on OSX for running the test:
/usr/bin/time -lp go test ./...
or on linux
/usr/bin/time -f "res kb %M" go test ./...

It should print useful output.

fixes #26